refactor(drops): centralize drop attribute fetching logic and simplify conditionals
